Added illustris-lhalo reader + fixed hdf5 save routines

+++ src/core_allvars.h
@@ -234,7 +234,7 @@ struct forest_info {
     // number of trees in each file because some files may have more/less trees whilst still spanning the
     // same volume (e.g., a void would contain few trees whilst a dense knot would contain many).
     int32_t *FileNr; // The file number that each forest was read from.
-    int32_t *original_treenr; // The tree number from the original tree files.
+    int64_t *original_treenr; // The tree number from the original tree files.
                               // Necessary because Task N's "Tree 0" could start at the middle of a file.
 };
